+++ drivers/misc/sprdwcn/platform/bufring.h
@@ -24,7 +24,7 @@
 #define MDBG_RING_LOCK_SIZE (sizeof(struct mutex))
 
 struct mdbg_ring_t {
-	long int size;
+	unsigned long int size;
 	char *pbuff;
 	char *rp;
 	char *wp;
@@ -36,14 +36,14 @@ struct mdbg_ring_t {
 	struct mutex *plock;
 };
 
-struct mdbg_ring_t *mdbg_ring_alloc(long int size);
+struct mdbg_ring_t *mdbg_ring_alloc(unsigned long int size);
 void mdbg_ring_destroy(struct mdbg_ring_t *ring);
 int mdbg_ring_read(struct mdbg_ring_t *ring, void *buf, int len);
 int mdbg_ring_write(struct mdbg_ring_t *ring, void *buf, unsigned int len);
 int mdbg_ring_write_timeout(struct mdbg_ring_t *ring, void *buf,
 			    unsigned int len, unsigned int timeout);
 char *mdbg_ring_write_ext(struct mdbg_ring_t *ring, long int len);
-bool mdbg_ring_will_full(struct mdbg_ring_t *ring, int len);
+bool mdbg_ring_will_full(struct mdbg_ring_t *ring, long int len);
 long int mdbg_ring_free_space(struct mdbg_ring_t *ring);
 long int mdbg_ring_readable_len(struct mdbg_ring_t *ring);
 void mdbg_ring_clear(struct mdbg_ring_t *ring);
